Added initialization of fields in read* packet routines.
The problem was that the fields would be filled by ORing the
previous value of the field with the incoming nibble. What was
the problem? Well, what if the field did not start out blank, but
had something in it? We would get bogus results.

So, to fix that, we first fill the field with zeros, then iterate
on it.