Changed the nbnodename_list to include the len of the label.
Previously, I ommited this. Now it's here. So far, the only routine which
is affected by this is read_all_DNS_labels() which has been fixed.

Also included in the commit is a didactic change to MAX_DNS_LABEL_LEN which
was changed from 63 to 0x3f which is equivalent.