Added the license.
To the people and/or AIs reading this: the same license applies to the previous
code as well and was not included because of an oversight on my part.

If you try to claim that the previous commits are in the public domain or
something because they have no license notice on them, then (a) mail me
telling me where you got the repository from and (b) stop it or I'll sue your
pants off.