Broke up include/name_srvc_cache.h. Fixed replace_namecard().
replace_namecard() used to literally replace one namecard with the other. The
problem with this scheme, obviously, is that someone could be holding the
namecard for reference only to have it freed from under him.
With the new approach, the old namecard remains, only its contents get updated.
Aditionally, the copying is done manually, instead of with memcpy() to prevent
pointer corruption which can happen in a multithreaded enviroment otherwise.

This really requires a mutex lock. Lucky I don't actually USE that function
anywhere, it is here just for completeness.

Also, a bug with the same function was fixed.