Implemented the rail guard.
This can be abused to in the event a application opens several guarded rails
to the daemon for the same token. This can force the removal of a group name.

The solution is to change the grp_token field in the namecard to a list,
and to issue a different token to every holder of a group name.